Prevent overflows when calculating on page fields
Using the original size is nice when displaying the data but makes it
more likely to have an overflow on an interim calculation when using
them for logic. Better use unsigned rather than uint8_t or uint16_t.

+++ include/parse_log_sense.h
@@ -101,7 +101,7 @@ inline static uint8_t log_sense_param_fmt(uint8_t *param)
 	return log_sense_param_flags(param) & LOG_PARAM_FLAG_FMT_MASK;
 }
 
-static inline uint8_t log_sense_param_len(uint8_t *param)
+static inline unsigned log_sense_param_len(uint8_t *param)
 {
 	return param[3];
 }
